7-Zip Version 2.30 Beta 18
- SDK/Archive/Cab/MSZipDecoder.cpp
  SDK/Archive/Cab/LZXDecoder.cpp:
    bug with corrupted archives was fixed
- Project/Compress/LZ/MatchFinder/BinTree/BinTree.h
- Project/Compress/LZ/MatchFinder/BinTree/BinTreeMain.h
    some speed optimization (using prefetching)

+++ Project/Archiver/Format/7z/Handler.h
@@ -53,7 +53,7 @@ END_COM_MAP()
 DECLARE_NOT_AGGREGATABLE(CHandler)
 
 DECLARE_REGISTRY(CHandler, TEXT("SevenZip.Format7z.1"), 
-                 TEXT("SevenZip.Format7z"), 0, THREADFLAGS_APARTMENT)
+                 TEXT("SevenZip.Format7z"), UINT(0), THREADFLAGS_APARTMENT)
 
   STDMETHOD(Open)(IInStream *aStream, 
       const UINT64 *aMaxCheckStartPosition,
